Add support for pausing synchronization
Motivation
------------
During cluster maintenance it may be desirable to pause all
synchronization for an index set.

Modifications
---------------
Add a paused setting to the CRD which defaults to false,
and disable sync jobs whenever it is set to true. Ensure that
we still run cleanup on delete.

Results
--------
Users may choose to temporarily disable the synchronization
process when desired without deleting the indices are turning
off the operator completely.

Fixes #16